Test package only depends on vim25
This means it can be used for testing throughout the derivative packages
without the possibility of circular dependencies.
